+++ gd-search-common/src/main/java/de/seitenbau/govdata/search/index/model/HitDto.java
@@ -53,6 +53,8 @@ public class HitDto
 
   private Boolean hasClosed;
 
+  private Boolean hasDataService;
+
   private String ownerOrg;
 
   private List<ResourceDto> resources;